+++ assets/javascripts/kitten/components/form/dropdown-select/index.js
@@ -44,8 +44,6 @@ export const DropdownSelect = ({ combobox, ...props }) => {
 
   const itemToString = item => (item ? String(item.label) : '')
 
-  const initialSelectedItem = find(['value', defaultSelectedValue])(options)
-
   const onSelectedItemChange = changes => {
     onChange(changes.selectedItem)
     onInputChange({ value: changes.selectedItem })
@@ -53,7 +51,7 @@ export const DropdownSelect = ({ combobox, ...props }) => {
 
   // Turns a hierarchy of options with `children` into a flat array
   // of options with a `level` of 1, 2 or null.
-  const flattenedOptions = () => {
+  const flattenedOptions = (function () {
     const flatOptions = []
 
     options.map(option => {
@@ -69,9 +67,11 @@ export const DropdownSelect = ({ combobox, ...props }) => {
         flatOptions.push(option)
       }
     })
-
     return flatOptions
-  }
+  })()
+  const initialSelectedItem = find(['value', defaultSelectedValue])(
+    flattenedOptions,
+  )
 
   const onIsOpenChange = changes => {
     if (changes.isOpen) return onMenuOpen({ changes })
@@ -90,7 +90,7 @@ export const DropdownSelect = ({ combobox, ...props }) => {
   } = useSelect({
     id: `${id}_element`,
     toggleButtonId: id,
-    items: flattenedOptions(),
+    items: flattenedOptions,
     getA11ySelectionMessage,
     itemToString,
     initialSelectedItem,
@@ -168,7 +168,7 @@ export const DropdownSelect = ({ combobox, ...props }) => {
       </button>
       <ul className="k-Form-Dropdown__list" {...getMenuProps()}>
         {isOpen &&
-          flattenedOptions().map((item, index) => (
+          flattenedOptions.map((item, index) => (
             <li
               className={classNames(
                 'k-Form-Dropdown__item',