+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `noMargin` prop to `FieldInput` component.
+- Fix: Set default value for `Dropdown` options on level 2
 
 ## [3.1.0] - 2021-04-08